Разделил на модули main.js.
