+++ js/util.js
@@ -1,0 +1,11 @@
+const getRandomValue = (min, max) => {
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomValue(0, elements.length - 1)];
+
+export {getRandomValue, getRandomArrayElement};
